stm32/rtc: move lowpower stuff to a separate mod.
